+++ lib/schedules/schedules.dart
@@ -3,4 +3,5 @@ export 'data/const/schedules_const.dart';
 export 'data/data_providers/schedules_data_provider.dart';
 export 'data/repositories/schedules_repository.dart';
 export 'view/schedules_page.dart';
-export 'widgets/day_schedules.dart';
+export 'widgets/day_schedules_item.dart';
+export 'widgets/day_schedules_list.dart';
